# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI-Assisted Developer Productivity Platform API
  description: |
    API for managing AI-assisted development workflows, tracking AI tool usage,
    orchestrating code agents, and integrating with CI/CD pipelines.
  version: 1.0.0
  contact:
    name: API Support
    email: support@ai-dev-platform.example
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://ai-dev-platform-staging.example.com/api/v1
    description: Staging server
  - url: https://ai-dev-platform.example.com/api/v1
    description: Production server
...
paths:
...
  /ai-activities:
    get:
      tags: [AI Activities]
      summary: List AI activities
      description: Retrieve a paginated list of AI activities
      operationId: listAIActivities
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - in: query
          name: project_id
          schema:
            type: string
            format: uuid
          description: Filter by project
        - in: query
          name: tool_used
          schema:
            type: string
            enum: [chatgpt, claude, copilot, cursor]
          description: Filter by AI tool
        - in: query
          name: category
          schema:
            type: string
            enum: [feature, bugfix, refactor, docs, test]
          description: Filter by category
      responses:
        '200':
          description: AI activities retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AIActivity'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [AI Activities]
      summary: Log AI activity
      description: Record an AI tool interaction
      operationId: logAIActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIActivityCreate'
      responses:
        '201':
          description: AI activity logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIActivity'
  /ai-activities/{activity_id}:
    get:
      tags: [AI Activities]
      summary: Get AI activity
      description: Retrieve a specific AI activity
      operationId: getAIActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityIdParam'
      responses:
        '200':
          description: AI activity retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIActivity'
  /agents:
    post:
      tags: [Agents]
      summary: Execute agent task
      description: Orchestrate a code agent to perform a task
      operationId: executeAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecutionCreate'
      responses:
        '201':
          description: Agent task initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentExecution'
  /agents/{execution_id}:
    get:
      tags: [Agents]
      summary: Get agent execution
      description: Retrieve status and results of an agent execution
      operationId: getAgentExecution
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Agent execution retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentExecution'
  /agents/types:
    get:
      tags: [Agents]
      summary: List available agent types
      description: Retrieve all available code agent types
      operationId: listAgentTypes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agent types retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentType'
...
  /analytics/usage:
    get:
      tags: [Analytics]
      summary: Get usage analytics
      description: Retrieve AI tool usage statistics and analytics
      operationId: getUsageAnalytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: project_id
          schema:
            type: string
            format: uuid
          description: Filter by project
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          description: Start date for analytics period
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          description: End date for analytics period
      responses:
        '200':
          description: Usage analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UsageAnalytics'
  /analytics/productivity:
    get:
      tags: [Analytics]
      summary: Get productivity metrics
      description: Retrieve developer productivity metrics
      operationId: getProductivityMetrics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: project_id
          schema:
            type: string
            format: uuid
          description: Filter by project
      responses:
        '200':
          description: Productivity metrics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProductivityMetrics'
  /mcp/servers:
    get:
      tags: [MCP]
      summary: List MCP servers
      description: Retrieve all configured MCP servers
      operationId: listMCPServers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: MCP servers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MCPServer'

    post:
      tags: [MCP]
      summary: Register MCP server
      description: Register a new MCP server
      operationId: registerMCPServer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPServerCreate'
      responses:
        '201':
          description: MCP server registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MCPServer'
  /mcp/servers/{server_id}/execute:
    post:
      tags: [MCP]
      summary: Execute MCP tool
      description: Execute a tool on an MCP server
      operationId: executeMCPTool
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_name, parameters]
              properties:
                tool_name:
                  type: string
                parameters:
                  type: object
      responses:
        '200':
          description: MCP tool executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    PerPageParam:
      in: query
      name: per_page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
...
    ActivityIdParam:
      in: path
      name: activity_id
      required: true
      schema:
        type: string
        format: uuid
      description: Activity UUID
    ExecutionIdParam:
      in: path
      name: execution_id
      required: true
      schema:
        type: string
        format: uuid
      description: Execution UUID
    ServerIdParam:
      in: path
      name: server_id
      required: true
      schema:
        type: string
        format: uuid
      description: Server UUID
  schemas:
...
    AIActivity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        tool_used:
          type: string
          enum: [chatgpt, claude, copilot, cursor]
        prompt:
          type: string
        response:
          type: string
        code_changes:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        category:
          type: string
          enum: [feature, bugfix, refactor, docs, test]
    AIActivityCreate:
      type: object
      required: [project_id, tool_used, prompt, category]
      properties:
        project_id:
          type: string
          format: uuid
        tool_used:
          type: string
          enum: [chatgpt, claude, copilot, cursor]
        prompt:
          type: string
        response:
          type: string
        code_changes:
          type: array
          items:
            type: string
        category:
          type: string
          enum: [feature, bugfix, refactor, docs, test]
    AgentExecution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        agent_type:
          type: string
        task_description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        input_data:
          type: object
        output_data:
          type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
    AgentExecutionCreate:
      type: object
      required: [project_id, agent_type, task_description]
      properties:
        project_id:
          type: string
          format: uuid
        agent_type:
          type: string
        task_description:
          type: string
        input_data:
          type: object
    AgentType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        requires_mcp_server:
          type: boolean
...
    UsageAnalytics:
      type: object
      properties:
        total_prompts:
          type: integer
        prompts_by_tool:
          type: object
          additionalProperties:
            type: integer
        prompts_by_category:
          type: object
          additionalProperties:
            type: integer
        total_cost_estimate:
          type: number
          format: float
        avg_tokens_per_prompt:
          type: number
          format: float
        time_saved_hours:
          type: number
          format: float
    ProductivityMetrics:
      type: object
      properties:
        total_commits:
          type: integer
        lines_of_code_changed:
          type: integer
        ai_assisted_commits:
          type: integer
        ai_contribution_percentage:
          type: number
          format: float
        test_coverage:
          type: number
          format: float
        avg_build_time_minutes:
          type: number
          format: float
    MCPServer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        server_type:
          type: string
          enum: [github, filesystem, database, http_api]
        endpoint:
          type: string
          format: uri
        status:
          type: string
          enum: [active, inactive, error]
        capabilities:
          type: array
          items:
            type: string
        last_health_check:
          type: string
          format: date-time
    MCPServerCreate:
      type: object
      required: [name, server_type, endpoint]
      properties:
        name:
          type: string
        server_type:
          type: string
          enum: [github, filesystem, database, http_api]
        endpoint:
          type: string
          format: uri
        capabilities:
          type: array
          items:
            type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer